requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
newsapi==0.1.1
newsapi-python==0.2.6
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
